Extracted keyboard handling out to FormVC
